Moved Scripts to normal Assets, began the lighting pass, attempted to make time move gradually.
