+++ .idea/.idea.SuperHotClone2018/.idea/contentModel.xml
@@ -1,12 +1,20 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project version="4">
   <component name="ContentModelStore">
-    <e p="C:\Users\seank\.Rider2018.2\config\extensions" t="IncludeRecursive" />
-    <e p="C:\Users\seank\.Rider2018.2\system\extResources" t="IncludeRecursive" />
-    <e p="C:\Users\seank\.Rider2018.2\system\resharper-host\Local\Transient\ReSharperHost\v182\SolutionCaches\_SuperHotClone2018.751243418.00" t="ExcludeRecursive" />
-    <e p="D:\int_game_dev\SuperHotClone2018" t="IncludeRecursive">
+    <e p="C:\Users\Ren Hughes\.Rider2018.2\system\extResources" t="IncludeRecursive" />
+    <e p="C:\Users\Ren Hughes\.Rider2018.2\system\resharper-host\Local\Transient\ReSharperHost\v182\SolutionCaches\_SuperHotClone2018.-710113824.00" t="ExcludeRecursive" />
+    <e p="C:\Users\Ren Hughes\Documents\NYU\SuperHotClone2018" t="IncludeRecursive">
       <e p="Assembly-CSharp.csproj" t="IncludeRecursive" />
       <e p="Assets" t="Include">
+        <e p="Charlie's Scripts" t="Include">
+          <e p="ProjectileMove.cs" t="Include" />
+          <e p="RotateToMouse.cs" t="Include" />
+          <e p="SpawnProjectiles.cs" t="Include" />
+        </e>
+        <e p="scriptsRen" t="Include">
+          <e p="cubeProtoMiove.cs" t="Include" />
+          <e p="TimeManager.cs" t="Include" />
+        </e>
         <e p="scriptsSean" t="Include">
           <e p="itemPickUp.cs" t="Include" />
           <e p="lookPickUp.cs" t="Include" />
