+++ Assets/Prefabs/PrefabsRen.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 74171de8a69051b42b79fba08803b671
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
