+++ Assets/Prefabs/PrefabsRen/ControllerMove.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 77c948335ee513d4bb34d28a33377c58
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
