+++ Assets/Prefabs/PrefabsRen/TimeManager.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 9e8590def9fa0f6449d9fdc698ee5e14
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
