+++ Assets/Prefabs/PrefabsRen/exampleCube.prefab.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 31635176bdc08a745a90f414ad128157
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 100100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
