+++ Assets/TestMaterials/MaterialsRen.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e4b7675124a0a904882fe84ffa55fc95
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
