+++ Assets/TestMaterials/MaterialsRen/blueclear.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 4857f80af39d6454b98b939d23b96f03
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
