+++ Assets/TestMaterials/MaterialsRen/red.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 36c18436b7625a24e870908b89009b83
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
