+++ Assets/scriptsRen/TimeManager.cs
@@ -8,13 +8,12 @@ public class TimeManager : MonoBehaviour
 {
 
 
-	//public float Delta;
+	//public float slowAcceleration = 0.3f;
+	//public float slowIncriment = 0.2f; 
 
-	//public float FixedDelta;
-
-	//public float TimeScale = 1;
-
-	public float slowDownFactor = 0.5f; //sets variable for on a scale from 0=>1, how slow do we go? 
+	public float slowDownFactor; //sets variable for on a scale from 0=>1, how slow do we go? 
+	
+	public float slowDownLength = 2f; 
 
 	//public float slowdownLength;
 
@@ -25,33 +24,14 @@ public class TimeManager : MonoBehaviour
 		instance = this; //setting the static instance
 	}
 
-	// Update is called once per frame
-	void Update () {
-		/*if (Input.GetKeyDown(KeyCode.P))
-		{
-		//	if (Time.timeScale == 1.0f)
-
-			//TimeScale = 0.5f; 
-			
-			//else
-			
-			//	Time.timeScale = 1.0f; 
-			
-		SlowDown();
-			
-			//Time.fixedDeltaTime = 0.02f * Time.timeScale; 
-		}
-		*/
-		
-		
-
-		//Delta = Time.deltaTime * TimeScale;
-	}
 
 	public void SlowDown() //a function to call when you want something to slow down
 	{
+		
+		
 		Time.timeScale = slowDownFactor; //set the time scale = to the slowdownfactor
-		Debug.Log("TimeMan"); //debugging to be sure it's reading it
+		Time.fixedDeltaTime = Time.timeScale * 0.02f; 
+		Debug.Log(Time.timeScale); //debugging to be sure it's reading it
 	}
