+++ Assets/scriptsRen/cubeProtoMiove.cs.meta
